+++ src/api/java/betterwithmods/common/blocks/mechanical/tile/TileEntityWaterwheel.java
@@ -1,0 +1,7 @@
+package betterwithmods.common.blocks.mechanical.tile;
+
+import net.minecraft.util.math.BlockPos;
+
+public abstract class TileEntityWaterwheel extends TileAxleGenerator {
+    public abstract boolean isWater(BlockPos pos);
+}