+++ src/api/java/betterwithmods/util/DirUtils.java
@@ -1,0 +1,8 @@
+package betterwithmods.util;
+
+import net.minecraft.block.properties.PropertyEnum;
+import net.minecraft.util.EnumFacing;
+
+public abstract class DirUtils {
+    public static final PropertyEnum<EnumFacing.Axis> AXIS = null;
+}