+++ src/main/scala/streams/StreamsMod.scala
@@ -28,6 +28,7 @@ object StreamsMod extends FarseekBaseMod {
         if(event.getSide == CLIENT) {
             info("Welcome to Farseek/Streams. Please visit http://www.minecraftforum.net/forums/mapping-and-modding/minecraft-mods/2346379-streams-real-flowing-rivers for information and updates.")
             info("Shader configuration: if you wish to use shaders with Streams river blocks, put the block.properties file from this mod's jar into your shaderpack's 'shaders' directory.")
+            info("Note: if shaders stop working for Streams blocks after a restart, disable and re-enable your shaderpack to fix this.")
         }
     }
 }